# Reduce total memory allocated and number of allocations in PubSub processing