--- a/pubsub_test.go
+++ b/pubsub_test.go
@@ -312,6 +312,22 @@ func TestPubSubChaotic(t *T) {
 	}
 }
 
+func BenchmarkPubSub(b *B) {
+	c, pubC := PubSub(dial()), dial()
+	msg := string(randStr())
+	msgCh := make(chan PubSubMessage, 1)
+	require.Nil(b, c.Subscribe(msgCh, "benchmark"))
+
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		if err := pubC.Do(Cmd(nil, "PUBLISH", "benchmark", msg)); err != nil {
+			b.Fatal(err)
+		}
+		<-msgCh
+	}
+}
+
 func ExamplePubSub() {
 	// Create a normal redis connection
 	conn, err := Dial("tcp", "127.0.0.1:6379")