Also includes a new benchmark for simple PubSub

```
name      old time/op    new time/op    delta
PubSub-8    51.9µs ±10%    46.0µs ±14%  -11.24%  (p=0.002 n=10+10)

name      old alloc/op   new alloc/op   delta
PubSub-8    4.73kB ± 0%    0.37kB ± 0%  -92.18%  (p=0.000 n=10+10)

name      old allocs/op  new allocs/op  delta
PubSub-8      17.0 ± 0%      10.0 ± 0%  -41.18%  (p=0.000 n=10+10)
```

Updates #91