@mediocregopher 2019-04-24
Thanks justin! We could probably save some more time by not using RawMessage at all, but that can be done another time, since this already saves so much in allocations as it is.